refactor: merge model and data directory
