--- app/src/main/java/de/uniks/codecamp/group_a/weather/data/source/remote/parsing/Wind.kt
@@ -1,5 +1,0 @@
-package de.uniks.codecamp.group_a.weather.data.source.remote.parsing
-
-data class Wind(
-    val speed: Double,
-)
