+++ app/src/main/java/de/uniks/codecamp/group_a/weather/injection/AppModule.kt
@@ -7,7 +7,7 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import de.uniks.codecamp.group_a.weather.data.source.remote.ApiService
+import de.uniks.codecamp.group_a.weather.model.source.remote.ApiService
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.create
