+++ app/src/main/java/de/uniks/codecamp/group_a/weather/injection/RepositoryModule.kt
@@ -4,7 +4,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import de.uniks.codecamp.group_a.weather.data.repository.WeatherRepositoryImpl
+import de.uniks.codecamp.group_a.weather.model.repository.WeatherRepositoryImpl
 import de.uniks.codecamp.group_a.weather.domain.repository.WeatherRepositoryInterface
 import javax.inject.Singleton
