+++ app/src/main/java/de/uniks/codecamp/group_a/weather/model/WeatherData.kt
@@ -1,7 +1,13 @@
 package de.uniks.codecamp.group_a.weather.model
 
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "weather")
 data class WeatherData(
+    @PrimaryKey
     val time: String,
+
     val temperature: Double,
     val location: String?, // e.g. "Kassel"
     val description: String, // e.g. "Partly Cloudy"
