+++ app/src/main/java/de/uniks/codecamp/group_a/weather/model/source/remote/parsing/WeatherDescription.kt
@@ -1,0 +1,5 @@
+package de.uniks.codecamp.group_a.weather.model.source.remote.parsing
+
+data class WeatherDescription(
+    val description: String,
+)
