+++ app/src/main/java/de/uniks/codecamp/group_a/weather/model/source/remote/parsing/Wind.kt
@@ -1,0 +1,5 @@
+package de.uniks.codecamp.group_a.weather.model.source.remote.parsing
+
+data class Wind(
+    val speed: Double,
+)
